more edits
The new table contained several undefined quantities or was referencing
BH only. I have corrected this. I have also restored the phase/amplitude
group naming convnetion example. I further moved the definition of the
coordinate separation to the frame section and referenced the Eq in the
Table. I have expanded thedescriptions in the table to be absolutely
clear.